pyproject.toml
[tool.poetry]
name = "instagram_clone"
version = "0.1.0"
description = ""
authors = ["Daniel S. Waite <danielwaite190@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
django-phone-field = "^1.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
